Add move, connection animation

+++ app/src/main/java/ua/mykhailenko/maps/model/City.kt
@@ -1,6 +1,6 @@
 package ua.mykhailenko.maps.model
 
 data class City(var name: String, var longitude: Float, var latitude: Float) {
-    var x: Double = 0.0
-    var y: Double = 0.0
+    var x: Float = 0.0f
+    var y: Float = 0.0f
 }
